requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pymongo==4.6.0
motor==3.3.2
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
spacy==3.7.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aiosmtplib==3.0.1
jinja2==3.1.2